🔧 Change package name to eslint-config-joyn
Using scoped package names for shareable ESLint config has some
limitations, see
https://eslint.org/docs/developer-guide/shareable-configs#sharing-multiple-configs.

There are two naming patterns for shareable ESLint config package names:

1. `eslint-config-${name}`
2. `@${name}/eslint/config`

It is always possible to reference them in a ESLint config with the full
package name, but when it comes to using the concise names the second
version has some limitations.
Extending from the base config works the some for both versions:

1. `extends: [ 'joyn' ]`
2. `extends: [ '@joyn' ]`

But when extending a "subconfig" the second version doesn’t allow the
use of the concise name:

1. `extends: [ 'joyn/presets/node' ]`
2. `extends: [ '@joyn/presets/node ']` doesn’t work, but `extends: [ '@joyn/eslint-config/presets/node' ]` works

Therefore I think it is easiert to only use the first version of the package name.